delete old cache implementation
