+++ packages/houdini/runtime/cache/tests/list.test.ts
@@ -3,8 +3,345 @@ import { testConfig } from 'houdini-common'
 // locals
 import { Cache, rootID } from '../cache'
 import { SubscriptionSelection } from '../../types'
+import { RefetchUpdateMode } from '../../types'
 
 const config = testConfig()
+
+test('prepend linked lists update', function () {
+	// instantiate the cache
+	const cache = new Cache(config)
+
+	const selection = {
+		viewer: {
+			type: 'User',
+			keyRaw: 'viewer',
+			fields: {
+				id: {
+					type: 'ID',
+					keyRaw: 'id',
+				},
+				firstName: {
+					type: 'String',
+					keyRaw: 'firstName',
+				},
+				friends: {
+					type: 'User',
+					keyRaw: 'friends',
+					update: 'prepend',
+					fields: {
+						id: {
+							type: 'ID',
+							keyRaw: 'id',
+						},
+						firstName: {
+							type: 'String',
+							keyRaw: 'firstName',
+						},
+					},
+				},
+			},
+		},
+	} as SubscriptionSelection
+
+	// add some data to the cache
+	cache.write({
+		selection,
+		data: {
+			viewer: {
+				id: '1',
+				firstName: 'bob',
+				friends: [
+					{
+						id: '2',
+						firstName: 'jane',
+					},
+					{
+						id: '3',
+						firstName: 'mary',
+					},
+				],
+			},
+		},
+		applyUpdates: true,
+	})
+
+	// make sure we can get the linked lists back
+	expect(
+		cache.read({
+			selection: {
+				friends: {
+					type: 'User',
+					keyRaw: 'friends',
+					update: RefetchUpdateMode.prepend,
+					fields: {
+						id: {
+							type: 'ID',
+							keyRaw: 'id',
+						},
+						firstName: {
+							type: 'String',
+							keyRaw: 'firstName',
+						},
+					},
+				},
+			},
+			parent: 'User:1',
+		})
+	).toEqual({
+		friends: [
+			{
+				id: '2',
+				firstName: 'jane',
+			},
+			{
+				id: '3',
+				firstName: 'mary',
+			},
+		],
+	})
+	// add some data to the cache
+	cache.write({
+		selection,
+		data: {
+			viewer: {
+				id: '1',
+				firstName: 'bob',
+				friends: [
+					{
+						id: '4',
+						firstName: 'jane',
+					},
+					{
+						id: '5',
+						firstName: 'mary',
+					},
+				],
+			},
+		},
+		applyUpdates: true,
+	})
+
+	// make sure we can get the linked lists back
+	expect(
+		cache.read({
+			selection: {
+				friends: {
+					type: 'User',
+					keyRaw: 'friends',
+					update: RefetchUpdateMode.prepend,
+					fields: {
+						id: {
+							type: 'ID',
+							keyRaw: 'id',
+						},
+						firstName: {
+							type: 'String',
+							keyRaw: 'firstName',
+						},
+					},
+				},
+			},
+			parent: 'User:1',
+		})
+	).toEqual({
+		friends: [
+			{
+				id: '4',
+				firstName: 'jane',
+			},
+			{
+				id: '5',
+				firstName: 'mary',
+			},
+			{
+				id: '2',
+				firstName: 'jane',
+			},
+			{
+				id: '3',
+				firstName: 'mary',
+			},
+		],
+	})
+})
+
+test('append in list', function () {
+	// instantiate a cache
+	const cache = new Cache(config)
+
+	const selection = {
+		viewer: {
+			type: 'User',
+			keyRaw: 'viewer',
+			fields: {
+				id: {
+					type: 'ID',
+					keyRaw: 'id',
+				},
+				friends: {
+					type: 'User',
+					keyRaw: 'friends',
+					list: {
+						name: 'All_Users',
+						connection: false,
+						type: 'User',
+					},
+					fields: {
+						id: {
+							type: 'ID',
+							keyRaw: 'id',
+						},
+						firstName: {
+							type: 'String',
+							keyRaw: 'firstName',
+						},
+					},
+				},
+			},
+		},
+	}
+
+	// start off associated with one object
+	cache.write({
+		selection,
+		data: {
+			viewer: {
+				id: '1',
+				friends: [
+					{
+						id: '2',
+						firstName: 'jane',
+					},
+				],
+			},
+		},
+	})
+
+	// a function to spy on that will play the role of set
+	const set = jest.fn()
+
+	// subscribe to the fields
+	cache.subscribe({
+		rootType: 'Query',
+		set,
+		selection,
+	})
+
+	// insert an element into the list (no parent ID)
+	cache.list('All_Users').append(
+		{ id: { type: 'ID', keyRaw: 'id' }, firstName: { type: 'String', keyRaw: 'firstName' } },
+		{
+			id: '3',
+			firstName: 'mary',
+		}
+	)
+
+	// make sure we got the new value
+	expect(set).toHaveBeenCalledWith({
+		viewer: {
+			id: '1',
+			friends: [
+				{
+					firstName: 'jane',
+					id: '2',
+				},
+				{
+					firstName: 'mary',
+					id: '3',
+				},
+			],
+		},
+	})
+})
+
+test('prepend in list', function () {
+	// instantiate a cache
+	const cache = new Cache(config)
+
+	const selection = {
+		viewer: {
+			type: 'User',
+			keyRaw: 'viewer',
+			fields: {
+				id: {
+					type: 'ID',
+					keyRaw: 'id',
+				},
+				friends: {
+					type: 'User',
+					keyRaw: 'friends',
+					list: {
+						name: 'All_Users',
+						connection: false,
+						type: 'User',
+					},
+					fields: {
+						id: {
+							type: 'ID',
+							keyRaw: 'id',
+						},
+						firstName: {
+							type: 'String',
+							keyRaw: 'firstName',
+						},
+					},
+				},
+			},
+		},
+	}
+
+	// start off associated with one object
+	cache.write({
+		selection,
+		data: {
+			viewer: {
+				id: '1',
+				friends: [
+					{
+						id: '2',
+						firstName: 'jane',
+					},
+				],
+			},
+		},
+	})
+
+	// a function to spy on that will play the role of set
+	const set = jest.fn()
+
+	// subscribe to the fields
+	cache.subscribe({
+		rootType: 'Query',
+		set,
+		selection,
+	})
+
+	// insert an element into the list (no parent ID)
+	cache.list('All_Users').prepend(
+		{ id: { type: 'ID', keyRaw: 'id' }, firstName: { type: 'String', keyRaw: 'firstName' } },
+		{
+			id: '3',
+			firstName: 'mary',
+		}
+	)
+
+	// make sure we got the new value
+	expect(set).toHaveBeenCalledWith({
+		viewer: {
+			id: '1',
+			friends: [
+				{
+					firstName: 'mary',
+					id: '3',
+				},
+				{
+					firstName: 'jane',
+					id: '2',
+				},
+			],
+		},
+	})
+})
 
 test('remove from connection', function () {
 	// instantiate a cache
@@ -122,15 +459,14 @@ test('remove from connection', function () {
 	})
 
 	// make sure we aren't subscribing to user 2 any more
-	expect(
-		cache.internal.getRecord(cache.id('User', '2')!)?.getSubscribers('firstName')
-	).toHaveLength(0)
+	expect(cache._internal_unstable.subscriptions.get('User:2', 'firstName')).toHaveLength(0)
 	// but we're still subscribing to user 3
+	expect(cache._internal_unstable.subscriptions.get('User:3', 'firstName')).toHaveLength(1)
+	// make sure we marked the corresponding edge for deletion
 	expect(
-		cache.internal.getRecord(cache.id('User', '3')!)?.getSubscribers('firstName')
-	).toHaveLength(1)
-	// make sure we deleted the edge holding onto this information
-	expect(cache.internal.getRecord('User:1.friends.edges[0]#User:2')).toBeUndefined()
+		cache._internal_unstable.storage.topLayer.operations['User:1.friends.edges[0]#User:2']
+			.deleted
+	).toBeTruthy()
 })
 
 test('append in connection', function () {
@@ -898,9 +1234,7 @@ test('remove from list', function () {
 	})
 
 	// make sure we aren't subscribing to user 2 any more
-	expect(
-		cache.internal.getRecord(cache.id('User', '2')!)?.getSubscribers('firstName')
-	).toHaveLength(0)
+	expect(cache._internal_unstable.subscriptions.get('User:2', 'firstName')).toHaveLength(0)
 })
 
 test('delete node', function () {
@@ -967,8 +1301,7 @@ test('delete node', function () {
 
 	// remove user 2 from the list
 	cache.delete(
-		'User',
-		cache.id('User', {
+		cache._internal_unstable.id('User', {
 			id: '2',
 		})!
 	)
@@ -982,7 +1315,7 @@ test('delete node', function () {
 	})
 
 	// make sure its empty now
-	expect(cache.internal.getRecord('User:2')).toBeFalsy()
+	expect(cache._internal_unstable.storage.topLayer.operations['User:2'].deleted).toBeTruthy()
 })
 
 test('delete node from connection', function () {
@@ -1071,8 +1404,7 @@ test('delete node from connection', function () {
 
 	// remove user 2 from the list
 	cache.delete(
-		'User',
-		cache.id('User', {
+		cache._internal_unstable.id('User', {
 			id: '2',
 		})!
 	)
@@ -1088,7 +1420,7 @@ test('delete node from connection', function () {
 	})
 
 	// make sure its empty now
-	expect(cache.internal.getRecord('User:2')).toBeFalsy()
+	expect(cache._internal_unstable.storage.topLayer.operations['User:2'].deleted).toBeTruthy()
 })
 
 test('append operation', function () {
@@ -1141,7 +1473,7 @@ test('append operation', function () {
 					},
 				},
 			},
-			parentID: cache.id('User', '1')!,
+			parentID: cache._internal_unstable.id('User', '1')!,
 			set: jest.fn(),
 		},
 		{}
@@ -1160,7 +1492,7 @@ test('append operation', function () {
 						list: 'All_Users',
 						parentID: {
 							kind: 'String',
-							value: cache.id('User', '1')!,
+							value: cache._internal_unstable.id('User', '1')!,
 						},
 					},
 				],
@@ -1180,7 +1512,7 @@ test('append operation', function () {
 	})
 
 	// make sure we just added to the list
-	expect([...cache.list('All_Users', cache.id('User', '1')!)]).toHaveLength(1)
+	expect([...cache.list('All_Users', cache._internal_unstable.id('User', '1')!)]).toHaveLength(1)
 })
 
 test('append from list', function () {
@@ -1233,7 +1565,7 @@ test('append from list', function () {
 					},
 				},
 			},
-			parentID: cache.id('User', '1')!,
+			parentID: cache._internal_unstable.id('User', '1')!,
 			set: jest.fn(),
 		},
 		{}
@@ -1252,7 +1584,7 @@ test('append from list', function () {
 						list: 'All_Users',
 						parentID: {
 							kind: 'String',
-							value: cache.id('User', '1')!,
+							value: cache._internal_unstable.id('User', '1')!,
 						},
 					},
 				],
@@ -1270,7 +1602,7 @@ test('append from list', function () {
 	})
 
 	// make sure we just added to the list
-	expect([...cache.list('All_Users', cache.id('User', '1')!)]).toHaveLength(2)
+	expect([...cache.list('All_Users', cache._internal_unstable.id('User', '1')!)]).toHaveLength(2)
 })
 
 test('append when operation', function () {
@@ -1329,7 +1661,7 @@ test('append when operation', function () {
 					},
 				},
 			},
-			parentID: cache.id('User', '1')!,
+			parentID: cache._internal_unstable.id('User', '1')!,
 			set: jest.fn(),
 		},
 		{}
@@ -1348,7 +1680,7 @@ test('append when operation', function () {
 						list: 'All_Users',
 						parentID: {
 							kind: 'String',
-							value: cache.id('User', '1')!,
+							value: cache._internal_unstable.id('User', '1')!,
 						},
 						when: {
 							must: {
@@ -1373,7 +1705,7 @@ test('append when operation', function () {
 	})
 
 	// make sure we just added to the list
-	expect([...cache.list('All_Users', cache.id('User', '1')!)]).toHaveLength(0)
+	expect([...cache.list('All_Users', cache._internal_unstable.id('User', '1')!)]).toHaveLength(0)
 })
 
 test('prepend when operation', function () {
@@ -1432,7 +1764,7 @@ test('prepend when operation', function () {
 					},
 				},
 			},
-			parentID: cache.id('User', '1')!,
+			parentID: cache._internal_unstable.id('User', '1')!,
 			set: jest.fn(),
 		},
 		{}
@@ -1451,7 +1783,7 @@ test('prepend when operation', function () {
 						list: 'All_Users',
 						parentID: {
 							kind: 'String',
-							value: cache.id('User', '1')!,
+							value: cache._internal_unstable.id('User', '1')!,
 						},
 						position: 'first',
 						when: {
@@ -1477,7 +1809,7 @@ test('prepend when operation', function () {
 	})
 
 	// make sure we just added to the list
-	expect([...cache.list('All_Users', cache.id('User', '1')!)]).toHaveLength(0)
+	expect([...cache.list('All_Users', cache._internal_unstable.id('User', '1')!)]).toHaveLength(0)
 })
 
 test('prepend operation', function () {
@@ -1550,7 +1882,7 @@ test('prepend operation', function () {
 					},
 				},
 			},
-			parentID: cache.id('User', '1')!,
+			parentID: cache._internal_unstable.id('User', '1')!,
 			set: jest.fn(),
 		},
 		{}
@@ -1569,7 +1901,7 @@ test('prepend operation', function () {
 						list: 'All_Users',
 						parentID: {
 							kind: 'String',
-							value: cache.id('User', '1')!,
+							value: cache._internal_unstable.id('User', '1')!,
 						},
 						position: 'first',
 					},
@@ -1590,9 +1922,10 @@ test('prepend operation', function () {
 	})
 
 	// make sure we just added to the list
-	expect(
-		[...cache.list('All_Users', cache.id('User', '1')!)].map((record) => record!.fields.id)
-	).toEqual(['3', '2'])
+	expect([...cache.list('All_Users', cache._internal_unstable.id('User', '1')!)]).toEqual([
+		'User:3',
+		'User:2',
+	])
 })
 
 test('remove operation', function () {
@@ -1660,7 +1993,7 @@ test('remove operation', function () {
 					},
 				},
 			},
-			parentID: cache.id('User', '1')!,
+			parentID: cache._internal_unstable.id('User', '1')!,
 			set: jest.fn(),
 		},
 		{}
@@ -1679,7 +2012,7 @@ test('remove operation', function () {
 						list: 'All_Users',
 						parentID: {
 							kind: 'String',
-							value: cache.id('User', '1')!,
+							value: cache._internal_unstable.id('User', '1')!,
 						},
 					},
 				],
@@ -1699,7 +2032,7 @@ test('remove operation', function () {
 	})
 
 	// make sure we removed the element from the list
-	expect([...cache.list('All_Users', cache.id('User', '1')!)]).toHaveLength(0)
+	expect([...cache.list('All_Users', cache._internal_unstable.id('User', '1')!)]).toHaveLength(0)
 })
 
 test('remove operation from list', function () {
@@ -1770,7 +2103,7 @@ test('remove operation from list', function () {
 					},
 				},
 			},
-			parentID: cache.id('User', '1')!,
+			parentID: cache._internal_unstable.id('User', '1')!,
 			set: jest.fn(),
 		},
 		{}
@@ -1789,7 +2122,7 @@ test('remove operation from list', function () {
 						list: 'All_Users',
 						parentID: {
 							kind: 'String',
-							value: cache.id('User', '1')!,
+							value: cache._internal_unstable.id('User', '1')!,
 						},
 					},
 				],
@@ -1807,7 +2140,7 @@ test('remove operation from list', function () {
 	})
 
 	// make sure we removed the element from the list
-	expect([...cache.list('All_Users', cache.id('User', '1')!)]).toHaveLength(0)
+	expect([...cache.list('All_Users', cache._internal_unstable.id('User', '1')!)]).toHaveLength(0)
 })
 
 test('delete operation', function () {
@@ -1875,7 +2208,7 @@ test('delete operation', function () {
 					},
 				},
 			},
-			parentID: cache.id('User', '1')!,
+			parentID: cache._internal_unstable.id('User', '1')!,
 			set: jest.fn(),
 		},
 		{}
@@ -1910,9 +2243,9 @@ test('delete operation', function () {
 	})
 
 	// make sure we removed the element from the list
-	expect([...cache.list('All_Users', cache.id('User', '1')!)]).toHaveLength(0)
+	expect([...cache.list('All_Users', cache._internal_unstable.id('User', '1')!)]).toHaveLength(0)
 
-	expect(cache.internal.getRecord('User:2')).toBeFalsy()
+	expect(cache._internal_unstable.storage.topLayer.operations['User:2'].deleted).toBeTruthy()
 })
 
 test('delete operation from list', function () {
@@ -1983,7 +2316,7 @@ test('delete operation from list', function () {
 					},
 				},
 			},
-			parentID: cache.id('User', '1')!,
+			parentID: cache._internal_unstable.id('User', '1')!,
 			set: jest.fn(),
 		},
 		{}
@@ -2018,10 +2351,10 @@ test('delete operation from list', function () {
 	})
 
 	// make sure we removed the element from the list
-	expect([...cache.list('All_Users', cache.id('User', '1')!)]).toHaveLength(0)
+	expect([...cache.list('All_Users', cache._internal_unstable.id('User', '1')!)]).toHaveLength(0)
 
-	expect(cache.internal.getRecord('User:2')).toBeFalsy()
-	expect(cache.internal.getRecord('User:3')).toBeFalsy()
+	expect(cache._internal_unstable.storage.topLayer.operations['User:2'].deleted).toBeTruthy()
+	expect(cache._internal_unstable.storage.topLayer.operations['User:3'].deleted).toBeTruthy()
 })
 
 test('disabled linked lists update', function () {
@@ -2083,20 +2416,22 @@ test('disabled linked lists update', function () {
 
 	// make sure we can get the linked lists back
 	expect(
-		cache.internal
-			.getRecord(cache.id('User', '1')!)
-			?.flatLinkedList('friends')
-			.map((data) => data!.fields)
-	).toEqual([
-		{
-			id: '2',
-			firstName: 'jane',
-		},
-		{
-			id: '3',
-			firstName: 'mary',
-		},
-	])
+		cache.read({
+			selection: { friends: selection.viewer.fields.friends },
+			parent: 'User:1',
+		})
+	).toEqual({
+		friends: [
+			{
+				id: '2',
+				firstName: 'jane',
+			},
+			{
+				id: '3',
+				firstName: 'mary',
+			},
+		],
+	})
 
 	// add some data to the cache
 	cache.write({
@@ -2121,20 +2456,22 @@ test('disabled linked lists update', function () {
 
 	// make sure we can get the linked lists back
 	expect(
-		cache.internal
-			.getRecord(cache.id('User', '1')!)
-			?.flatLinkedList('friends')
-			.map((data) => data!.fields)
-	).toEqual([
-		{
-			id: '3',
-			firstName: 'jane',
-		},
-		{
-			id: '4',
-			firstName: 'mary',
-		},
-	])
+		cache.read({
+			selection: { friends: selection.viewer.fields.friends },
+			parent: 'User:1',
+		})
+	).toEqual({
+		friends: [
+			{
+				id: '3',
+				firstName: 'jane',
+			},
+			{
+				id: '4',
+				firstName: 'mary',
+			},
+		],
+	})
 })
 
 test('append linked lists update', function () {
@@ -2196,20 +2533,22 @@ test('append linked lists update', function () {
 
 	// make sure we can get the linked lists back
 	expect(
-		cache.internal
-			.getRecord(cache.id('User', '1')!)
-			?.flatLinkedList('friends')
-			.map((data) => data!.fields)
-	).toEqual([
-		{
-			id: '2',
-			firstName: 'jane',
-		},
-		{
-			id: '3',
-			firstName: 'mary',
-		},
-	])
+		cache.read({
+			selection: { friends: selection.viewer.fields.friends },
+			parent: 'User:1',
+		})
+	).toEqual({
+		friends: [
+			{
+				id: '2',
+				firstName: 'jane',
+			},
+			{
+				id: '3',
+				firstName: 'mary',
+			},
+		],
+	})
 
 	// add some data to the cache
 	cache.write({
@@ -2235,28 +2574,30 @@ test('append linked lists update', function () {
 
 	// make sure we can get the linked lists back
 	expect(
-		cache.internal
-			.getRecord(cache.id('User', '1')!)
-			?.flatLinkedList('friends')
-			.map((data) => data!.fields)
-	).toEqual([
-		{
-			id: '2',
-			firstName: 'jane',
-		},
-		{
-			id: '3',
-			firstName: 'mary',
-		},
-		{
-			id: '4',
-			firstName: 'jane',
-		},
-		{
-			id: '5',
-			firstName: 'mary',
-		},
-	])
+		cache.read({
+			selection: { friends: selection.viewer.fields.friends },
+			parent: 'User:1',
+		})
+	).toEqual({
+		friends: [
+			{
+				id: '2',
+				firstName: 'jane',
+			},
+			{
+				id: '3',
+				firstName: 'mary',
+			},
+			{
+				id: '4',
+				firstName: 'jane',
+			},
+			{
+				id: '5',
+				firstName: 'mary',
+			},
+		],
+	})
 })
 
 test('writing a scalar marked with a disabled update overwrites', function () {
@@ -2298,7 +2639,7 @@ test('writing a scalar marked with a disabled update overwrites', function () {
 	})
 
 	// make sure we can get the linked lists back
-	expect(cache.internal.getData(cache.internal.record(rootID), selection, {})).toEqual({
+	expect(cache.read({ selection })).toEqual({
 		viewer: {
 			id: '1',
 			firstName: 'bob',
@@ -2319,7 +2660,7 @@ test('writing a scalar marked with a disabled update overwrites', function () {
 	})
 
 	// make sure we can get the updated lists back
-	expect(cache.internal.getData(cache.internal.record(rootID), selection, {})).toEqual({
+	expect(cache.read({ selection })).toEqual({
 		viewer: {
 			id: '1',
 			firstName: 'bob',
@@ -2367,7 +2708,7 @@ test('writing a scalar marked with a prepend', function () {
 	})
 
 	// make sure we can get the linked lists back
-	expect(cache.internal.getData(cache.internal.record(rootID), selection, {})).toEqual({
+	expect(cache.read({ selection })).toEqual({
 		viewer: {
 			id: '1',
 			firstName: 'bob',
@@ -2389,7 +2730,7 @@ test('writing a scalar marked with a prepend', function () {
 	})
 
 	// make sure we can get the updated lists back
-	expect(cache.internal.getData(cache.internal.record(rootID), selection, {})).toEqual({
+	expect(cache.read({ selection })).toEqual({
 		viewer: {
 			id: '1',
 			firstName: 'bob',
@@ -2437,7 +2778,7 @@ test('writing a scalar marked with an append', function () {
 	})
 
 	// make sure we can get the linked lists back
-	expect(cache.internal.getData(cache.internal.record(rootID), selection, {})).toEqual({
+	expect(cache.read({ selection })).toEqual({
 		viewer: {
 			id: '1',
 			firstName: 'bob',
@@ -2459,7 +2800,7 @@ test('writing a scalar marked with an append', function () {
 	})
 
 	// make sure we can get the updated lists back
-	expect(cache.internal.getData(cache.internal.record(rootID), selection, {})).toEqual({
+	expect(cache.read({ selection })).toEqual({
 		viewer: {
 			id: '1',
 			firstName: 'bob',
